flatten workspace into single binary crate
Flattens the workspace into a single binary crate by merging the
functionality of the `codec` crate into the rpc-perf::codec module.
