+++ Cargo.toml
@@ -1,15 +1,47 @@
-[workspace]
-members = [
-    "codec",
-    "rpc-perf",
-]
+[package]
+name = "rpc-perf"
+version = "3.2.0-pre"
+authors = ["Brian Martin <bmartin@twitter.com>"]
+edition = "2018"
+license = "Apache-2.0"
+description = "RPC Performance Testing"
+homepage = "https://github.com/twitter/rpc-perf"
+repository = "https://github.com/twitter/rpc-perf"
+readme = "README.md"
 
-[profile.bench]
-debug = true
-lto = true
-codegen-units = 1
+[dependencies]
+byteorder = "1.3.4"
+bytes = "0.6.0"
+clap = "2.33.3"
+crc = "1.8.1"
+mio = { version = "0.7.6", features = ["net", "os-poll"] }
+rand = "0.7.3"
+rustcommon-atomics = { git = "https://github.com/twitter/rustcommon", branch = "master" }
+rustcommon-buffer = { git = "https://github.com/twitter/rustcommon", branch = "master" }
+rustcommon-heatmap = { git = "https://github.com/twitter/rustcommon", branch = "master" }
+rustcommon-logger = { git = "https://github.com/twitter/rustcommon", branch = "master" }
+rustcommon-metrics = { git = "https://github.com/twitter/rustcommon", branch = "master" }
+rustcommon-ratelimiter = { git = "https://github.com/twitter/rustcommon", branch = "master" }
+rustcommon-timer = { git = "https://github.com/twitter/rustcommon", branch = "master" }
+rustcommon-waterfall = { git = "https://github.com/twitter/rustcommon", branch = "master" }
+rustls = { version = "0.18.1", features = ["dangerous_configuration"] }
+serde = "1.0.116"
+serde_derive = "1.0.116"
+slab = "0.4.2"
+strum = "0.19.2"
+strum_macros = "0.19.2"
+tiny_http = "0.7.0"
+toml = "0.5.6"
+webpki = "0.21.3"
 
 [profile.release]
+opt-level = 3
 debug = true
+rpath = false
 lto = true
+debug-assertions = false
 codegen-units = 1
+
+[profile.dev]
+debug = true
+opt-level = 0
