--- codec/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "codec"
-version = "0.3.0"
-authors = ["Brian Martin <bmartin@twitter.com>"]
-edition = "2018"
-license = "Apache-2.0"
-description = "Codecs and generators"
-homepage = "https://github.com/twitter/rpc-perf"
-repository = "https://github.com/twitter/rpc-perf"
-publish = false
-
-[dependencies]
-bytes = "0.6.0"
-crc = "1.8.1"
-rand = "0.7.3"
