+++ configs/thrift_cache_hash.toml
