+++ src/codec/memcache.rs
@@ -1,19 +1,23 @@
-// Copyright 2019 Twitter, Inc.
+// Copyright 2019-2020 Twitter, Inc.
 // Licensed under the Apache License, Version 2.0
 // http://www.apache.org/licenses/LICENSE-2.0
 
-use super::*;
+use crate::codec::*;
+use crate::config::Action;
+use crate::stats::Stat;
+use std::io::{BufRead, BufReader};
 
 use bytes::{Buf, BytesMut};
 
-use std::io::{BufRead, BufReader};
-
-#[derive(Default)]
-pub struct Memcache {}
+pub struct Memcache {
+    common: Common,
+}
 
 impl Memcache {
     pub fn new() -> Self {
-        Self {}
+        Self {
+            common: Common::new(),
+        }
     }
 
     pub fn get(&self, buf: &mut BytesMut, key: &[u8]) {
@@ -48,7 +52,21 @@ impl Memcache {
     }
 }
 
-impl Decoder for Memcache {
+impl Default for Memcache {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+impl Codec for Memcache {
+    fn common(&self) -> &Common {
+        &self.common
+    }
+
+    fn common_mut(&mut self) -> &mut Common {
+        &mut self.common
+    }
+
     fn decode(&self, buf: &[u8]) -> Result<Response, Error> {
         // Shortest response is "OK\r\n" at 4bytes
         if buf.len() < 4 {
@@ -168,6 +186,33 @@ impl Decoder for Memcache {
         }
 
         Err(Error::Unknown)
+    }
+
+    fn encode(&mut self, buf: &mut BytesMut, rng: &mut ThreadRng) {
+        let command = self.generate(rng);
+        match command.action() {
+            Action::Get => {
+                let key = command.key().unwrap();
+                if let Some(metrics) = self.common.metrics() {
+                    metrics.increment(&Stat::CommandsGet);
+                    metrics.distribution(&Stat::KeySize, key.len() as u64);
+                }
+                self.get(buf, key);
+            }
+            Action::Set => {
+                let key = command.key().unwrap();
+                let value = command.value().unwrap();
+                if let Some(metrics) = self.common.metrics() {
+                    metrics.increment(&Stat::CommandsSet);
+                    metrics.distribution(&Stat::KeySize, key.len() as u64);
+                    metrics.distribution(&Stat::ValueSize, value.len() as u64);
+                }
+                self.set(buf, key, value, command.ttl().map(|ttl| ttl as u32), None);
+            }
+            action => {
+                fatal!("Action: {:?} unsupported for Memcache", action);
+            }
+        }
     }
 }
