+++ src/stats/snapshot.rs
